feat: enhance image processing with improved negative control strategies and region handling
