+++ src/readii/loaders.py
@@ -16,147 +16,149 @@ import SimpleITK as sitk
 from imgtools import io
 from imgtools.ops import StructureSetToSegmentation
 
-from readii.utils import logger 
+from readii.utils import logger
+
 
 def loadDicomSITK(imgDirPath: str | Path) -> sitk.Image:
-    """Read a DICOM series as a SimpleITK Image.
+	"""Read a DICOM series as a SimpleITK Image.
 
-    Parameters
-    ----------
-        imgDirPath (Union[str, Path]): The path to the directory containing the DICOM series to
-        load. It can be either a string or a Path object.
+	Parameters
+	----------
+	    imgDirPath (Union[str, Path]): The path to the directory containing the DICOM series to
+	    load. It can be either a string or a Path object.
 
-    Returns
-    -------
-        sitk.Image: The loaded image.
-    """
-    # Convert to Path if passed as a string
-    imgDirPath = Path(imgDirPath)
+	Returns
+	-------
+	    sitk.Image: The loaded image.
+	"""
+	# Convert to Path if passed as a string
+	imgDirPath = Path(imgDirPath)
 
-    logger.debug(f"Loading DICOM series from directory: {imgDirPath}")
-    reader = sitk.ImageSeriesReader()
-    dicomNames = reader.GetGDCMSeriesFileNames(imgDirPath)
-    reader.SetFileNames(dicomNames)
-    return reader.Execute()
+	logger.debug(f"Loading DICOM series from directory: {imgDirPath}")
+	reader = sitk.ImageSeriesReader()
+	dicomNames = reader.GetGDCMSeriesFileNames(imgDirPath)
+	reader.SetFileNames(dicomNames)
+	return reader.Execute()
 
 
 def loadRTSTRUCTSITK(
-    rtstructPath: str | Path, baseImageDirPath: str | Path, roiNames: Optional[str] = None
+	rtstructPath: str | Path, baseImageDirPath: str | Path, roiNames: Optional[str] = None
 ) -> Dict[str, sitk.Image]:
-    """Load RTSTRUCT into SimpleITK Image.
+	"""Load RTSTRUCT into SimpleITK Image.
 
-    Parameters
-    ----------
-    rtstructPath : str or Path
-        Path to the DICOM file containing the RTSTRUCT.
-    baseImageDirPath : str or Path
-        Path to the directory containing the DICOMs for the original image the segmentation
-        was created from (e.g., CT). This is required to load the RTSTRUCT.
-    roiNames : str, optional
-        Identifier for which region(s) of interest to load from the total segmentation file.
+	Parameters
+	----------
+	rtstructPath : str or Path
+	    Path to the DICOM file containing the RTSTRUCT.
+	baseImageDirPath : str or Path
+	    Path to the directory containing the DICOMs for the original image the segmentation
+	    was created from (e.g., CT). This is required to load the RTSTRUCT.
+	roiNames : str, optional
+	    Identifier for which region(s) of interest to load from the total segmentation file.
 
-    Returns
-    -------
-    dict
-        A dictionary of mask ROIs from the loaded RTSTRUCT image as
-        SimpleITK image objects.
-    """
-    # Convert to Path if passed as a string
-    rtstructPath = Path(rtstructPath)
-    baseImageDirPath = Path(baseImageDirPath)
+	Returns
+	-------
+	dict
+	    A dictionary of mask ROIs from the loaded RTSTRUCT image as
+	    SimpleITK image objects.
+	"""
+	# Convert to Path if passed as a string
+	rtstructPath = Path(rtstructPath)
+	baseImageDirPath = Path(baseImageDirPath)
 
-    logger.debug(f"Loading RTSTRUCT from directory: {rtstructPath}")
-    baseImage: io.Scan = io.read_dicom_scan(baseImageDirPath.resolve())
-    segImage: io.StructureSet = io.read_dicom_rtstruct(rtstructPath.resolve())
+	logger.debug(f"Loading RTSTRUCT from directory: {rtstructPath}")
+	baseImage: io.Scan = io.read_dicom_scan(baseImageDirPath.resolve())
+	segImage: io.StructureSet = io.read_dicom_rtstruct(rtstructPath.resolve())
 
-    # Set up segmentation loader
-    logger.debug(f"Making mask using ROI names: {roiNames}")
-    makeMask = StructureSetToSegmentation(roi_names=roiNames)
+	# Set up segmentation loader
+	logger.debug(f"Making mask using ROI names: {roiNames}")
+	makeMask = StructureSetToSegmentation(roi_names=roiNames)
 
-    try:
-        # Get the individual ROI masks
-        segMasks = makeMask(
-            segImage,
-            baseImage.image,
-            existing_roi_indices={"background": 0},
-            ignore_missing_regex=False,
-        )
-    except ValueError:
-        return {}
+	try:
+		# Get the individual ROI masks
+		segMasks = makeMask(
+			segImage,
+			baseImage.image,
+			existing_roi_indices={"background": 0},
+			ignore_missing_regex=False,
+		)
+	except ValueError:
+		return {}
 
-    # Get list of ROIs present in this rtstruct
-    loadedROINames = segMasks.raw_roi_names
-    # Initialize dictionary to store ROI names and images
-    roiStructs = {}
-    # Get each roi and its label and store in dictionary
-    for roi in loadedROINames:
-        # Get the mask for this ROI
-        roiMask = segMasks.get_label(name=roi)
-        # Store the ROI name and image
-        roiStructs[roi] = roiMask
+	# Get list of ROIs present in this rtstruct
+	loadedROINames = segMasks.raw_roi_names
+	# Initialize dictionary to store ROI names and images
+	roiStructs = {}
+	# Get each roi and its label and store in dictionary
+	for roi in loadedROINames:
+		# Get the mask for this ROI
+		roiMask = segMasks.get_label(name=roi)
+		# Store the ROI name and image
+		roiStructs[roi] = roiMask
 
-    return roiStructs
+	return roiStructs
 
 
 def loadSegmentation(
-    segImagePath: str | Path,
-    modality: str,
-    baseImageDirPath: Optional[str | Path] = None,
-    roiNames: Optional[str] = None,
+	segImagePath: str | Path,
+	modality: str,
+	baseImageDirPath: Optional[str | Path] = None,
+	roiNames: Optional[str] = None,
 ) -> Dict[str, sitk.Image]:
-    """Load a segmentation with the correct function.
+	"""Load a segmentation with the correct function.
 
-    Parameters
-    ----------
-    segImagePath : str or Path
-        Path to the segmentation file to load.
-    modality : str
-        Type of image that segImagePath points to load. If RTSTRUCT, must set baseImageDirPath.
-    baseImageDirPath : str or Path, optional
-        Path to the directory containing the DICOMs for the original image the segmentation
-        was created from.
-    roiNames : str, optional
-        Identifier for which region(s) of interest to load from the total segmentation file.
+	Parameters
+	----------
+	segImagePath : str or Path
+	    Path to the segmentation file to load.
+	modality : str
+	    Type of image that segImagePath points to load. If RTSTRUCT, must set baseImageDirPath.
+	baseImageDirPath : str or Path, optional
+	    Path to the directory containing the DICOMs for the original image the segmentation
+	    was created from.
+	roiNames : str, optional
+	    Identifier for which region(s) of interest to load from the total segmentation file.
 
-    Returns
-    -------
-    dict
-        A dictionary of each of the ROIs and their name in the segmentation
-        image as sitk.Image objects.
+	Returns
+	-------
+	dict
+	    A dictionary of each of the ROIs and their name in the segmentation
+	    image as sitk.Image objects.
 
-    Examples
-    --------
-    >>> segImages = loadSegmentation(
-    ...     segImagePath="/path/to/segmentation/1.dcm",
-    ...     modality="RTSTRUCT",
-    ...     baseImageDirPath="/path/to/CT",
-    ...     roiNames="GTVp.*",
-    ... )
-    """
-    if modality.upper() not in ["SEG", "RTSTRUCT"]:
-        raise ValueError(
-            f"Unsupported segmentation modality '{modality}'. Must be one of 'RTSTRUCT' or 'SEG'."
-        )
-    # Always convert paths to Path objects
-    segImagePath = Path(segImagePath)
-    if baseImageDirPath is not None:
-        baseImageDirPath = Path(baseImageDirPath)
+	Examples
+	--------
+	>>> segImages = loadSegmentation(
+	...     segImagePath="/path/to/segmentation/1.dcm",
+	...     modality="RTSTRUCT",
+	...     baseImageDirPath="/path/to/CT",
+	...     roiNames="GTVp.*",
+	... )
+	"""
+	if modality.upper() not in ["SEG", "RTSTRUCT"]:
+		msg = f"Unsupported segmentation modality '{modality}'. Must be one of 'RTSTRUCT' or 'SEG'."
+		raise ValueError(msg)
 
-    if modality.upper() == "SEG":
-        imgFolder = segImagePath.parent
-        segHeader = pydicom.dcmread(segImagePath.resolve(), stop_before_pixels=True)
-        roiName = segHeader.SegmentSequence[0].SegmentLabel
-        return {roiName: loadDicomSITK(imgFolder)}
+	# Always convert paths to Path objects
+	segImagePath = Path(segImagePath)
 
-    # modality is RTSTRUCT
-    if baseImageDirPath is None:
-        raise ValueError(
-            "When loading RTSTRUCT, must pass baseImageDirPath since "
-            "RTSTRUCT loader requires original image."
-        )
+	if modality.upper() == "SEG":
+		imgFolder = segImagePath.parent
+		segHeader = pydicom.dcmread(segImagePath.resolve(), stop_before_pixels=True)
+		roiName = segHeader.SegmentSequence[0].SegmentLabel
+		return {roiName: loadDicomSITK(imgFolder)}
 
-    return loadRTSTRUCTSITK(
-        segImagePath.resolve(),
-        baseImageDirPath.resolve(),
-        roiNames,
-    )
+	# modality is RTSTRUCT
+	if baseImageDirPath is None:
+		msg = (
+			"When loading RTSTRUCT, must pass baseImageDirPath since "
+			"RTSTRUCT loader requires original image."
+		)
+		raise ValueError(msg)
+
+	baseImageDirPath = Path(baseImageDirPath)
+
+	return loadRTSTRUCTSITK(
+		segImagePath.resolve(),
+		baseImageDirPath.resolve(),
+		roiNames,
+	)
